refactor: tab name logic and heart and close button
