+++ _Frontend/src/main.tsx
@@ -11,4 +11,3 @@ createRoot(document.getElementById('root')!).render(
     </BrowserRouter>
   </StrictMode>,
 )
-// test
