+++ _Frontend/src/pages/Home/_components/Dashboard/TabSidebar/_hooks/tabConstants.ts
@@ -1,0 +1,5 @@
+export const TAB_COLORS = ['#ade3ef', '#adefb3', '#ddefad', '#fabcc5ff', '#dab7f9ff']
+
+export const BASE_TITLE = 'New Map'
+
+export const DEFAULT_TAB_COLOR = '#fffafa'
