+++ _Frontend/src/pages/Home/_components/Dashboard/TabSidebar/_hooks/types.ts
@@ -1,0 +1,5 @@
+export interface TabData {
+  id: string
+  title: string
+  color: string
+}
