+++ _Frontend/src/pages/Profile/_components/FavouriteComparisons.tsx
