+++ _Frontend/src/pages/Profile/_components/FavouriteLocations.tsx
